feat: 마이페이지 UI수정 및 page 생성
- 동행점수바 비율 조정, 사용자 별명 길이제한(최대8자)
- mypage/page.jsx(마이페이지 메인) 생성
- mypage/useredit/page.jsx(회원정보 수정페이지) 생성
- /mypage의 회원정보 수정 버튼 클릭 시 /useredit으로 이동 가능
- 기존컴포넌트(ProfileEdit,ProfileEditBgDt)삭제 및
  ProfileEdit(회원정보수정페이지 완성본)으로 통합
 - myj.html랑 test_myj.java도 삭제
