--- app/mypage/test_myj.java
@@ -1,1 +1,0 @@
-ajldjflajfdl;j
