+++ app/mypage/useredit/page.jsx
@@ -1,0 +1,12 @@
+"use client"
+
+import ProfileEdit from "@/components/mypage/ProfileEdit";
+
+export default function UserEdit() {
+
+  return (
+    <>
+      <ProfileEdit />
+    </>
+  );
+}
